Add tests verifying that the Google Tag Manager script gets injected on the page
